Added proper toString output and modified execute method
You want to return something in the toString method for when a error happens, you basically just recreate your syntax and use the variables, so if they provide null for something it will show what they provided null for. So if they did `set score at 5 of player to ""` with nothing in the quotes it will error saying `set score at 5 of player to null` so you know what's wrong. Also you should be null checking in the execute method that way you can send out the error rather that it printing a nullpointer.